Update Pelican and Markdown - add new plugin
A change in the Markdown package broke the YAML metadata parsing. I took
the opportunity to greatly simplify the way parsing the YAML metadata
was being done by writing a simple reader plugin.

This new plugin removes the need to have both a markdown extension to
parse the file, as well as shim to translate that data into something
that Pelican would expect to pull out of the markdown metadata.